split rendering and cursor positioning into their own functions, conform entirely to C89 standard

+++ structs.h
@@ -1,12 +1,10 @@
 #ifndef STRUCTS_H_
 #define STRUCTS_H_
 
-#include <stdint.h>
-
 /* structures for buffer and line editing */
 typedef struct {
     char* content;
-    uint64_t allocated;
+    unsigned int allocated;
 } Buffer;
 
 struct Line {
